requirements.txt
bson==0.5.8
ecdsa==0.13
flake8==3.7.7
imutils==0.5.1
mnemonic==0.18
mypy==0.701
opencv-python==3.4.3.18
prompt-toolkit==2.0.7
pyAesCrypt==0.4.2
pysha3==1.0.2
pytest-cov==2.6.1
pytest==4.4.0
python-bitcoinlib==0.10.1
pyyaml==3.13
pyzbar==0.1.7
qrcode[pil]==6.0
twine==1.13.0
sphinx==1.8.5
shamir-mnemonic==0.1
